Don't start wizard flow on atv when gms is present
This matches the behaviour prior to the LMODroid suw revamp on 21.
Setupwraith hooks things in a completely different manner to the mobile
gms suw. The LMODroid suw should not run at all when atv gms is present,
but the apk still needs to be available should the tvsetup intent hooks
get implemented.

Change-Id: Ic4943e8fd55d30fbd2856dc5eadc5a0ce6282f29